ci(tests): if tests failed the pipeline throw an error

+++ pipelines/cleanup/index.sh
@@ -3,8 +3,19 @@
 echo "Removing coverage files"
 ./find_workspace | awk '{print $1"/coverage"}' | xargs -l rm
 
-echo "Removing golangci-lint temp file"
-rm ./golangci-lint
+if [ -f ./golangci-lint ]
+then
+    echo "Removing golangci-lint temp file"
+    rm ./golangci-lint
+fi
 
-echo "Removing find_workspace temp file"
-rm ./find_workspace
+if [ -f ./find_workspace ]
+then
+    echo "Removing find_workspace temp file"
+    rm ./find_workspace
+fi
+
+if [ -f temp_test_results ]
+then
+    rm temp_test_results
+fi
